Contactform made into a component
