+++ src/components/Contact/_contact.scss
@@ -1,1 +1,2 @@
 @use '/src/styles/variables' as *;
+
